feat: add error handling infrastructure and architecture documentation
- Add errors.py with custom exception classes for validation and calculation
- Add test_errors.py with comprehensive error tests
- Update criterion.py with improved error handling
- Refactor plotting.py with better type handling
- Update calculate.py for consistency
- Add architecture review documentation and plan

+++ pyisomme/calculate.py
@@ -18,8 +18,8 @@ logger = logging.getLogger(__name__)
 
 @debug_logging(logger)
 def calculate_resultant(c1: Channel | None,
-                        c2: Channel | int | None = 0,
-                        c3: Channel | int | None = 0) -> Channel | None:
+                        c2: Channel | float | None = 0,
+                        c3: Channel | float | None = 0) -> Channel | None:
     """
     Takes 2 or 3 Channels and calculates the 2nd norm or resultant component.
     :param c1: X-Channel
